requirements.txt

langchain-qdrant
streamlit
requests
qdrant-client
langchain
langgraph
langsmith
google-generativeai
PyPDF2
pypdf
langchain-community